refactor: extract WebView features into dedicated component
- Create WebViewWithFeatures component to encapsulate dapp notification logic
- Move notifyChangedEventsToDappOnFocus from WebView props to features config
- Update PerpTrade to use WebViewWithFeatures with notification feature enabled
- Clean up WebView component by removing notification-specific code
