+++ packages/kit/src/views/Developer/pages/Gallery/Components/stories/PerpGallery.tsx
@@ -500,7 +500,7 @@ const PerpGallery = () => (
     componentName="PerpGallery"
     elements={[
       {
-        title: 'Hyperliquid API Test',
+        title: 'Hyperliquid API Test 2862',
         element: <PerpApiTests />,
       },
     ]}
